Uploaded the Hackathon Files
This is an app that we (team Exodia) built for the Dementia hackathon. The web app is a platform that allows family members of people with Dementia to set-up a schedule of automated audio reminders, phone calls, and more. Once a person with Dementia has completed a scheduled task, they press a combination of keys on their FireFly phone to verify task completion. The audio conversations with the Dementia patient are then available to researchers and scientists through the built-in REST api. We built this app in less than 32 hours and had a lot of fun doing so.